+++ programmable_video/lib/src/programmable_video.dart
@@ -141,4 +141,83 @@ class TwilioProgrammableVideo {
     }
     throw Exception('Permissions not granted');
   }
+
+  static Future<List<StatsReport>> getStats() async {
+    final statsMap = await ProgrammableVideoPlatform.instance.getStats();
+
+    return statsMap.entries.map((entry) {
+      final statReport = StatsReport(entry.key);
+      final values = entry.value;
+      values['localAudioTrackStats'].forEach((localAudioTrackStat) {
+        statReport.addLocalAudioTrackStats(LocalAudioTrackStats(
+          localAudioTrackStat['trackSide'],
+          localAudioTrackStat['packetsLost'],
+          localAudioTrackStat['codec'],
+          localAudioTrackStat['ssrc'],
+          localAudioTrackStat['timestamp'],
+          localAudioTrackStat['bytesSent'],
+          localAudioTrackStat['packetsSent'],
+          localAudioTrackStat['roundTripTime'],
+          localAudioTrackStat['audioLevel'],
+          localAudioTrackStat['jitter'],
+        ));
+      });
+
+      values['remoteAudioTrackStats'].forEach((remoteAudioTrackStat) {
+        statReport.addAudioTrackStats(RemoteAudioTrackStats(
+          remoteAudioTrackStat['trackSide'],
+          remoteAudioTrackStat['packetsLost'],
+          remoteAudioTrackStat['codec'],
+          remoteAudioTrackStat['ssrc'],
+          remoteAudioTrackStat['timestamp'],
+          remoteAudioTrackStat['bytesReceived'],
+          remoteAudioTrackStat['packetsReceived'],
+          remoteAudioTrackStat['audioLevel'],
+          remoteAudioTrackStat['jitter'],
+        ));
+      });
+
+      values['remoteAudioTrackStats'].forEach((remoteVideoTrackStat) {
+        statReport.addVideoTrackStats(RemoteVideoTrackStats(
+          remoteVideoTrackStat['trackSide'],
+          remoteVideoTrackStat['packetsLost'],
+          remoteVideoTrackStat['codec'],
+          remoteVideoTrackStat['ssrc'],
+          remoteVideoTrackStat['timestamp'],
+          remoteVideoTrackStat['bytesReceived'],
+          remoteVideoTrackStat['packetsReceived'],
+          VideoDimensions(
+            remoteVideoTrackStat['dimensionsHeight'],
+            remoteVideoTrackStat['dimensionsWidth'],
+          ),
+          remoteVideoTrackStat['frameRate'],
+        ));
+      });
+
+      values['localVideoTrackStats'].forEach((localVideoTrackStat) {
+        statReport.addLocalVideoTrackStats(LocalVideoTrackStats(
+          localVideoTrackStat['trackSide'],
+          localVideoTrackStat['packetsLost'],
+          localVideoTrackStat['codec'],
+          localVideoTrackStat['ssrc'],
+          localVideoTrackStat['timestamp'],
+          localVideoTrackStat['bytesSent'],
+          localVideoTrackStat['packetsSent'],
+          localVideoTrackStat['roundTripTime'],
+          VideoDimensions(
+            localVideoTrackStat['captureDimensionsHeight'],
+            localVideoTrackStat['captureDimensionsWidth'],
+          ),
+          VideoDimensions(
+            localVideoTrackStat['dimensionsHeight'],
+            localVideoTrackStat['dimensionsWidth'],
+          ),
+          localVideoTrackStat['capturedFrameRate'],
+          localVideoTrackStat['frameRate'],
+        ));
+      });
+
+      return statReport;
+    }).toList();
+  }
 }