+++ programmable_video/test/programmable_video_test.dart
@@ -58,6 +58,17 @@ void main() {
     });
   });
 
+  group('.getStats()', () {
+    final mockInterface = MockInterface();
+    setUpAll(() => ProgrammableVideoPlatform.instance = mockInterface);
+
+    test('should call interface code to get Stats', () async {
+      final result = await TwilioProgrammableVideo.getStats();
+      expect(mockInterface.getStatsWasCalled, true);
+      expect(result, []);
+    });
+  });
+
   group('.requestPermissionForCameraAndMicrophone()', () {
     test('should request camera and microphone permission', () async {
       var nativeRequestPermissionsIsCalled = false;