+++ programmable_video_platform_interface/lib/src/programmable_video_platform_interface.dart
@@ -63,6 +63,11 @@ abstract class ProgrammableVideoPlatform extends PlatformInterface {
     throw UnimplementedError('getSpeakerphoneOn() has not been implemented.');
   }
 
+  /// Calls native code to get stats
+  Future<Map<dynamic, dynamic>> getStats() {
+    throw UnimplementedError('getStats() has not been implemented.');
+  }
+
   Future<bool> deviceHasReceiver() {
     throw UnimplementedError('deviceHasReceiver() has not been implemented.');
   }