+++ programmable_video/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.7.2
+
+- Added method for fetching stats.
+
 ## 0.7.1
 
 - Added method for checking if device has a builtin earpiece.